+++ sumo/src/netimport/vissim/tempstructs/NIVissimExtendedEdgePoint.cpp
@@ -90,7 +90,7 @@ NIVissimExtendedEdgePoint::recheckLanes(const NBEdge * const edge) throw() {
     }
     // patch lane indices
     myLanes.clear();
-    for (int i=0; i<edge->getNoLanes(); ++i) {
+    for (int i=0; i<(int) edge->getNoLanes(); ++i) {
         myLanes.push_back(i);
     }
 }
